Adding persistence + dto <-> entity mapping with Mapstruct

+++ src/main/java/com/wefox/payments/dto/PaymentDto.java
@@ -3,8 +3,10 @@ package com.wefox.payments.dto;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.wefox.payments.enums.PaymentType;
+import com.wefox.payments.util.enums.PaymentType;
 import lombok.Data;
+
+import java.math.BigDecimal;
 
 @Data
 @JsonInclude(Include.NON_NULL)
@@ -22,6 +24,6 @@ public class PaymentDto {
   @JsonProperty("credit_card")
   private String creditCard;
 
-  private Integer amount;
+  private BigDecimal amount;
 
 }
